client/pyproject.toml
[project]
name = "fastApi-context"
version = "1.2.0"
description = "Auth check and more plugins for fastApi"
authors = [
    {name = "hxnie",email = "920521013@qq.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "starlette (>=0.46.2,<0.47.0)",
    "starlette-context (>=0.4.0,<0.5.0)",
    "pydantic (>=2.0,<3.0)",
]


[project.optional-dependencies]
redis = ["redis (>=5.2.1,<6.0.0)"]
jwt = ["python-jose (>=3.4.0,<4.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


license = "MIT"
keywords = ["fastapi", "context", "starlette", "middleware"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]